+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 * Added the option to disable `Autoplay` in the settings. 
 * Added formatting for some command outputs in the Discord chat.
 * Added a background scanner for the music files. You can now start the playback while the program is still scanning your library.
+* Added a fallback to stream via FFMpeg.
 * Updated the Discord.Net api to 2.0.1.
+* Switched from .Net Framework to .Net Core 2.1.
 
 ### Version 1.0.8 (2018-08-08)